select payments from current account

+++ db/models/accounts.js
@@ -83,7 +83,7 @@ const attributes = {
     field: "pending"
   },
   precision: {
-    type: DataTypes.DOUBLE,
+    type: DataTypes.INTEGER(11),
     allowNull: true,
     defaultValue: null,
     primaryKey: false,
